Reorganize contract host code
