+++ src/transactions/InvokeContractOpFrame.cpp
@@ -5,18 +5,13 @@
 #include "transactions/InvokeContractOpFrame.h"
 #include "ledger/LedgerTxn.h"
 #include "ledger/LedgerTxnEntry.h"
-#include "transactions/ContractHostTypes.h"
+#include "transactions/contracts/HostContext.h"
 #include "transactions/TransactionUtils.h"
 #include "util/GlobalChecks.h"
 #include "util/Logging.h"
 #include "xdr/Stellar-ledger-entries.h"
 #include "xdr/Stellar-transaction.h"
 #include <Tracy.hpp>
-#include <TracyC.h>
-#include <fizzy/execute.hpp>
-#include <fizzy/lib/fizzy/instantiate.hpp>
-#include <fizzy/parser.hpp>
-#include <fizzy/types.hpp>
 #include <memory>
 #include <stdexcept>
 #include <string_view>
