+++ src/transactions/TransactionFrame.cpp
@@ -19,7 +19,7 @@
 #include "ledger/LedgerTxnEntry.h"
 #include "ledger/LedgerTxnHeader.h"
 #include "main/Application.h"
-#include "transactions/ContractHostTypes.h"
+#include "transactions/contracts/HostContext.h"
 #include "transactions/SignatureChecker.h"
 #include "transactions/SignatureUtils.h"
 #include "transactions/SponsorshipUtils.h"
